TST: add render testing, mark as xslow
* This adds the first test that actually makes assertions about
the pixel content of `ggmolvis`-rendered images. It turns out
that regular render calls do "ok" in GitHub actions, but if you
use the compositor there seem to be issues. Since the test added here
uses the compositor to set the background color, it is marked `xslow`
so that the user/developer must opt in to run it.

* This approach has the advantage that we can start marking assertions
about the expected appearance of renders, but of course it has the
drawback that we can only run these tests locally for now (or using
one of the services that offers GPU CI runners, which some larger OSS
projects do, but usually has a cost). I think it is still better than
nothing given that we could at least bisect on newly-introduced problems
locally even if it is unfortunately rather easy for bugs to slip past
the CI this way. This is a common approach upstream for tests that are
not CI tractable (i.e., incredibly slow, or require unreasonably large
amounts of memory to reproduce the original problem).

* For a bit of a dissection on composition causing crashes in GHA CI,
see recent testing on my fork at:
https://github.com/tylerjereddy/ggmolvis/pull/1